Rework Docker part & Installation.
Use Docker in [swarm mode](https://docs.docker.com/engine/swarm/). It is
better to manage Docker services. In addition, the compose file services
are splitted so each one has is own role. One for webserver, one PHP (+ FPM), one for db & one for adminer.

A Makefile is added to easily run Docker services (containers)  and
install app in 3 commands.

A reverse-proxy is provided if the developer does not have one launched.
(see README).

Run the following commands to run the project :

- make start-reverse-proxy # If is needed.
- make start
- make install